pyproject.toml
[tool.poetry]
name = "a-trade"
version = "1.4.1"
description = ""
authors = ["fengjianpeng <miracle1627@163.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.8"
tushare = "^1.4.13"
python-dotenv = "^1.0.1"
matplotlib = "<3.9"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
pywencai = "^0.12.2"
jieba = "^0.42.1"
openai = "^1.54.3"
openpyxl = "^3.1.5"
wechatarticles = {path = "../wechat_articles_spider"}
selenium = "^4.27.1"
numpy = "^1.24.4"
opencv-python = "^4.5.5.64"
webdriver-manager = "^4.0.2"
volcengine-python-sdk = "^1.0.120"
json-repair = "0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
